delete tests dir
