+++ src/generate_image.rs
@@ -31,6 +31,7 @@ pub fn generate_image(prompt: &str) -> Result<String> {
         .ok_or(anyhow!("Failed to extract image URL"))?
         .to_string();
 
+    // log::info!("image_url {}", image_url);
     Ok(image_url)
 }
